backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
neo4j==5.15.0
redis==5.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP and async
httpx==0.25.2
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
pytz==2023.3